requirements.txt
# Зависимости для Telegram бота "Офисный ниндзя"

# Основная библиотека для Telegram Bot API
python-telegram-bot==20.7

# Для работы с переменными окружения
python-dotenv==1.0.0

# Дополнительные зависимости (опционально)
# requests==2.31.0  # если нужны HTTP запросы
# aiohttp==3.9.1    # если нужны асинхронные HTTP запросы

